Set Up Fuel Sliders
Set up sliders to indicated how much fuel is left for thrust and how
much of the cooldown is left on the boost. Also made minor values
changes to SpaceMove and increased regen speed on the ground. Also
readded a second planet. I'm probably going to make this a release and
then begin multiplayer.