Added Main Thruster Particle Effects (Networked!)
Unfortunately, the thruster effects aren't rocket flames that are in
world space and inherit velocity, since for whatever reason that
*particular* combination of settings results in jittery particles for
some reason. Instead, players just leave behind
still-relative-to-world-space orbs when they fire their thrusters.
Grenades have a similar trail, but it is always on (and disappears as
soon as the grenade does, which is ugly and will be fixed).